allow changing some simulation parameters

+++ source/EngineInterface/MonitorData.h
@@ -2,7 +2,7 @@
 
 struct MonitorData
 {
-    int timeStep = 0;
+    uint64_t timeStep = 0;
     int numCells = 0;
     int numParticles = 0;
     int numTokens = 0;
